Renames class
